Update patch set 7
Patch Set 7:

> Patch Set 7:
> > Switch order of leaf and app list menus on the right side so that
> > the leaf menu is the one on the screen edge.
> 
> Was that a wise decision? I think the Team menu should actually be in the screen corner.

The Leaf menu should go on the screen edge IMHO because it is the main menu. The team menu is used frequently but is unique to this mode. If we really wanted the Team menu to be the primary menu instead of the Leaf menu we'd have to switch the menu order on the left side instead!

> * I find it a bit hard to hit the knobbly widget ... It used to be at the screen edge and therefore easy to grab (-> Fitt's law). Maybe a changed mouse pointer ...

I agree that switching modes is a little bit tougher in new mini mode because the dragger knob is not on a screen edge but it's worth it because having quick access and spatial consistency of the leaf menu is more important. The resize knob has always been off the screen edge and it works fine.

The knobbly widget should be more visually distinct. If you look at old Deskbar the tray was visually recessed and the knob was on a separate light gray area but we've lost that somewhere along the way. I experimented with changing the mouse cursor when I implemented width resizing earlier, but, I ran into bugs and haven't gotten back to it. The appropriate action would be to change the mouse pointer on mouse down to indicate that you've entered dragging mode.

> * Can we remove the truncating of the app names in the team menu and just make the menu as wide as needed? As it is only shown when clicked, we don't need to conserve space like when having Deskbar in normal mode.

This is a good suggestion but the menu should be plenty wide enough to display any app name now. The width it reserves for the application name is the same as in horizontal mode and the width scales up with font and icon size.

> * Why is there no Expando mode for the team menu in mini mode?

That's step 3!

Patch-set: 7
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>