Update patch set 8
Patch Set 8:

> The Leaf menu should go on the screen edge IMHO because it is the main menu. The team menu is used frequently but is unique to this mode. If we really wanted the Team menu to be the primary menu instead of the Leaf menu we'd have to switch the menu order on the left side instead!

Yep. I'd opt for the increased ease to reach the Team menu instead of a "main" menu that is needed less often.

> > * I find it a bit hard to hit the knobbly widget ... It used to be at the screen edge and therefore easy to grab (-> Fitt's law). Maybe a changed mouse pointer ...

> The knobbly widget should be more visually distinct. [...] The appropriate action would be to change the mouse pointer on mouse down to indicate that you've entered dragging mode.

Agreed to having the knobblies more distinct. But mouse pointer change on mouse down isn't enough. It's nice to have a resizing pointer when doing the resizing moving (on all knobbly widgets, e.g. split view widgets as well). But to make it more discoverable, the mouse pointer should already change when it hovers above it. Then you know your aim is true and you're ready to drag/move with a mouse down.
> 
> > * Can we remove the truncating of the app names in the team menu and just make the menu as wide as needed? As it is only shown when clicked, we don't need to conserve space like when having Deskbar in normal mode.
> 
> This is a good suggestion but the menu should be plenty wide enough to display any app name now.

It's not wide enough for "LegacyPackageInstaller" for example.

> The width it reserves for the application name is the same as in horizontal mode and the width scales up with font and icon size.

That upscaling in horizontal mode makes sense, because you have to be conservative with the space. In vertical mode, we're restricted by the fixed width the user sets. But in mini-mode we could use any width (within reason), because the Team menu is only seen when clicked.

Patch-set: 8