Update patch set 10
Patch Set 10:

> Patch Set 8:
> I'd opt for the increased ease to reach the Team menu instead of a "main" menu that is needed less often.

It's the Apple menu, the Start menu, the Be menu, the Leaf menu, the unnameable menu that is always found in the corner of the screen somewhere and allows you to shut down your computer and maybe open an app or a preference. I'm trying to get used to the idea of the team menu on screen edge but I don't like it. Deskbar is based on the Mac System 7 menu bar with the Apple/Be menu on left and the app switcher on the right with the clock and other icons moved below (and now to the side.) On System 7 the menu bar was full screen so both menus were got a corner but we gotta pick one or the other - I vote leaf.

> But mouse pointer change on mouse down isn't enough. It's nice to have a resizing pointer when doing the resizing moving (on all knobbly widgets, e.g. split view widgets as well). But to make it more discoverable, the mouse pointer should already change when it hovers above it. Then you know your aim is true and you're ready to drag/move with a mouse down.

Fair enough. My main concern with your approach is that our state machine isn't currently good enough to reliably return your mouse pointer back to the hand cursor in every single case so we might get stuck with the wrong mouse pointer icon. Changing the cursor on mouse down would make that a lot more rare with worst-case that you'd only need an additional click to fix it.

> > > * Can we remove the truncating of the app names in the team menu and just make the menu as wide as needed?

> It's not wide enough for "LegacyPackageInstaller" for example.

> That upscaling in horizontal mode makes sense, because you have to be conservative with the space. In vertical mode, we're restricted by the fixed width the user sets. But in mini-mode we could use any width (within reason), because the Team menu is only seen when clicked.

The problem is almost the same as in horizontal mode because in horizontal mode the menu items can be as wide as they need to be until we hit the status tray and then we have to start truncating. So the app width that is picked in horizontal mode should be wide enough to accommodate most app names.

There is a constant in Deskbar called gMinimumWindowWidth which sets the minimum window width in vertical expando mode. This is the width that we use for application names. This same width is used in horizontal mode as the default app item width and in the team menu in mini-mode as well.

In the last patch and this one I took that width, used since the days of yore and scaled it based on the current font and icon size so that you have about the same amount of room to fit the icon and application name. This is illustrated pretty well here: https://insightfactory.tumblr.com/image/190905126687
The top image shows the items at the default gMininimumWindowWidth while the bottom image shows items scaled to fit the bigger font and icon size so that you can fit the same amount of text in the box at the bigger font size.

In this patch the same width and scaling are used as are used in horizontal mode in the last patch. If your saying that "LegacyPackageInstaller" doesn't fit then what you're saying is that gMinimumWindowWidth should be increased a bit so that the name will fit in horizontal mode and vertical expando mode (at least at the default 12pt font) too.

Team menu is a bit different than those other modes in that it is a floating vertical menu so we could add a bit more space here but there is still a balancing act between too little space for names to fit and too much space for you to comfortably move your mouse to the submenu so we want to keep the width within reason. I hope that this gives you a better appreciation for why I picked the width that I did even though it might seem arbitrary.

> Patch Set 8:
> 
> Why Team_Registrar icon is 64x64, but actual content is 16x16?

The Team_Registrar icon is a vector hvif icon that is native 64x64. The actual bitmap used is scaled and rasterized from the vector icon. The icon size is actually 19x19 which is the same size as the bitmap it replaces (actually 19x15.) The bitmap size is then scaled up with (bold) font size so that you increase your font size the menu gets bigger and the icon gets bigger too.

Patch-set: 10
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>