Update patch set 2
Patch Set 2:

> Patch Set 2:
> Invalid files were created then testing #15585 and entering KDL manually. Maybe some timeouts are expired and something went wrong.

That's something that should be looked into.

> Checkfs does not detect some error types (#15609) and can delete valid files. 

I just proof-read the sources, and yes, it currently cannot handle the case when the block cache runs out of memory; I even designed the API in a way that makes this impossible, as the actual error code is not returned.

It's not a huge change, though, and should definitely be handled gracefully by checkfs.

Patch-set: 2
Reviewer: Gerrit User 1000016 <1000016@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>