Update patch set 4
Patch Set 4:

> Does it have to break the API?

Not for the two functions you mentionned, but it would still remove UrlEncode and UrlDecode, or at least, that's the point of the change: URLs should always be built with the correct parameters, and then the user of BUrl should not worry about wether a BUrl object was already url-encoded or not.

Also, the default for both the constructor and SetUrlString is to encode the URL, which was not the case before. So, even if the ABI could be preserved, the API would still change, probably leading to more subtle breakage.

Patch-set: 4
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>