Update patch set 3
Patch Set 3: Code-Review-2

Changing kernel args breaks the bootloader compatibility vs. older versions of Haiku. This will require users to maintain 2 EFI loaders if they want to boot beta1, or any version up till this point. We should try not to do this now that we are in beta.

Could we change kernel args to be a KMessage, maybe? Then we could break things once and be fine after that. But still, breaking things in this way is not good.

Patch-set: 3
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-2