Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Thanks for the review Adrien!
> 
> > Patch Set 1: Code-Review+2
> > 
> > I think you can create a BCertificate::Private from X509 data in openssl format and then create a BCertificate from that.
> > 
> > The expected use case for the API was that you would add the exception using the certificate from CertificateValidationFailed when it's called for the first time (after the user has checked that it's indeed valid and requested the exgeption to be r membered)
> 
> Yeah I saw that, I just didn't really want to depend on anything except for the public interface in these tests to avoid having them break from some internal refactor.
> 
> I think what I want to do, for now at least, is add some test cases to HttpsTest which just uses a different listener to conditionally add an exception so that we can at least have some regression tests to show that certificate verification happens.

I started adding some additional tests:

1. Verify that certificate validation fails for self-signed certificate.
2. Verify that, via listener callback, a certificate exception can be added,
3. Verify that, once a certificate exception is added, using a different hostname which resolves to the same server, but doesn't match the common-name field in the server's certificate, fails.

I think test case #3 exposes a bug here. My assumption was that adding a certificate exception meant that the certificate was now trusted, but that other validation would still be performed. Now that I look at the code more carefully it seems that this is not the actual behavior.

The certificate exception just means that any use of this certificate is now trusted, even if someone else is using it. That means that if I tell the system to trust certificate from Server A which was not already trusted by the system's CA, and someone who owns Server B gets a hold of the certificate, then Server B will also be trusted. That seems wrong to me.

I think I want to continue this in a followup patch in order to prevent the current one from becoming too enormous, but that's up to you obviously.

Patch-set: 1
Reviewer: Gerrit User 1000162 <1000162@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>