Update patch set 13
Patch Set 13:

> What do you bean by "both sides"? If 2 team buttons, than this is weird. Layout should be same as in compact mode.

The team menu is on the left when deskbar is on the left edge of the screen, and on the right when deskbar is on the right edge of the screen.

As for splitting bugs from features, I agree in general, but with DeskBar code not being super clean, everything tends to be a bit tangled together and it's sometimes hard to split changes apart.

Patch-set: 13