Update patch set 13
Patch Set 13:

> Patch Set 13:
> 
> > Patch Set 13: -Code-Review
> > Leaf menu height increase in vertical mode is on purpose, it matches the tab height.
> 
> There are no need in matching tab height in default mode. When auto-expand mode is enabled, height space is important.

It's not strictly needed no but it is much easier if the height of the menu bar height matches between mini and expendo mode. It's 4px taller which is accounted for by the bigger font 12pt vs 10pt font on BeOS. If the height is not increased you'd see an odd-looking 4px space between mini-mode deskbar and a maximized window.

> > Deskbar shelf icons should not be clipped, what are your plain and bold font settings?
> 
> I use "Noto Sans CJK JP" size 12 as plain and bold font. Icons should be not clipped no matter what font is selected.

I tested on a wide range of font settings and the only time there are overlapping items is when you have a very small bold font size and a very large plain font size which in practice should never happen. At 12pt Noto Sans if icons are clipped then there's a bug but I am perplexed how that is possible.

Patch-set: 13
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0