Update patch set 4
Patch Set 4:

I think the best solution here is to intentionally change behavior *and* keep BeOS compatibility -- i.e. do a "__gABIVersion" check and set a flag manually for <= ABI_GCC2_BEOS, as is already done for B_EXECUTE_AREA & B_CLONEABLE_AREA in create_area().

I'm not sure what use the old behavior has, though. I guess if ViewBitmaps worked properly, it would be irrelevant, as the background would never be drawn -- unless the bitmap was a format that had an alpha channel...?

Probably the flag should be inverted compared to the way it is in this version of the patch. That is, set the flag to get the "old" behavior (something like "B_DONT_DRAW_BACKGROUND", but that is kind of a gnarly name.)

Patch-set: 4
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>