Update patch set 1
Patch Set 1:

According to http://forth.sourceforge.net/wordset/16-color-text-ext/16colord1_2.pdf , the "foreground-color" and "background-color" user words are not necessarily available ("as User Interface commands if the FCode
Debugging command group of the User Interface is present.")

And the terminal-emulator package should handle the usual escape sequences, so we should be able to just use that.

However before using this we should check the "iso6429-1983-colors" of the terminal-emulator package to verify that the feature is available. If neither this nor the user words to change colors are available, we should try to display our menu in a readable way using only black and white, I guess.

The other changes (#columns vs screen-#columns, etc) should be reviewed more carefully and matched against the specs where possible. And moved to arch-specific places if needed, rather than having this file full of #ifdefs.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>