Deskbar: Reimplement auto-raise/implement auto-lower
This changes auto-raise mode so that when the mouse moves over Deskbar
it raises Deskbar up, and when the mouse leaves Deskbar it lowers it
back down again as described in #13304.

Always activate Deskbar on click, ignoring auto-raise setting.
Deskbar stays active as long as a menu is open.

Disable auto-raise in horizontal top mode and grey out the option in
Deskbar prefs. Deskbar can't work in horizontal top mode like the
other modes because if it did Deskbar would float on top of window tabs
making them inaccessible. Only activate Deskbar while a menu is open
in this mode, otherwise keep Deskbar behind all other windows.

Create an fBarWindow member in BarView::AttachedToWindow() and use it
throughout the class so that we don't have to keep getting it over and
over again. Change bool fIsMenuShown to int32 fMenusShown and use it to
store a reference count of open menus. In the current design menus
could be opened from multiple locations clobbering the bool.

Change order of methods in BarView to MouseDown(), MouseMoved(), then
MouseUp() because that order is both logical and alphabetical.

Don't raise or lower Deskbar if window has been dragged from outside.

Fixes #15678, #13304

Change-Id: I076a452b26250d9eb9a4eccb4a6aa6f939e11d34