Deskbar: Reimplement auto-raise/implement auto-lower
This changes auto-raise so that when the mouse moves over the screen
edge within the Deskbar window it raises Deskbar up and when the mouse
leaves Deskbar it lowers it back down again as described in #13304.

Raise/Activate Deskbar on click unless a menu has just been opened. If a
menu has been opened Deskbar is raised/activated by the window instead.
RaiseDeskbar() in auto-raise mode or Activate() in normal mode.
Activate() makes Deskbar the active window, RaiseDeskbar() does not.

Raise/Activate Deskbar on menu open. Deskbar stays on top as long as a
menu remains open. Lower Deskbar again on menu close in auto-raise mode so
that Deskbar goes back into the background.

Don't raise/lower Deskbar if window has been dragged from the outside
in auto-raise mode.

Change bool fShowingMenu to int32 fMenusShown and use it to store a
reference count of open menus. In the previous design menus could be
opened from multiple locations clobbering the bool.

Add an fBarWindow member to BarView and initialize it in AttachedToWindow()
Use this throughout the class so that we don't have to keep getting it over
and over again. Also add an fBarApp member to BarWindow and use that
instead of creating it again and again.

Change order of methods in BarView to MouseDown(), MouseMoved(), then
MouseUp() because that order is both logical and alphabetical.

Fixes #15678, #13304

Change-Id: I076a452b26250d9eb9a4eccb4a6aa6f939e11d34