Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> I tested it a bit some more with Web+ and its buttons on the right can't be reached anymore...
> Not sure what the best thing to do here. Maybe an option to only auto-raise when cursor is over leaf button + shelf?

This is precisely why I said that DeskBar should only raise when the mouse hits the screen edge, not whenever it is hovering the area where deskbar is. That area is too wide and there is no point allowing windows to use it if it's not possible to click that area.

Patch-set: 7