Update patch set 7
Patch Set 7:

> I think raising on hitting the screen edge is way too undiscoverable. Leaf button (+shelf) seems to be more obvious. We can rename the option to "Auto-rise on hovering leaf" or something to make it clearer?

It's very easy to have the cursor at the screen edge when you try to reach deskbar (that's the application of Fitt's law). It allows to raise deskbar in more places:
- By moving the cursor to the top edge of where deskbar would be, if you aim for the leaf menu,
- But also by moving the cursor to the screen side, if you aim for the team list

The latter is quite nice, it allows to implement some kind of gesture where you put the mouse cursor completely to the screen edge, and the deskbar (with team list) pops up and you can easily switch to other apps. This would not work if we only raised while over the leaf menu.

We can possibly have both: the whole leaf menu area + the right screen edge ?

I still fear the leaf menu would be too large of a target and result in accidental raise of the deskbar, however. Maybe it's small enough and in a not too much used corner, but I'm not sure, I'd have to try and see if it triggers accidentally a lot.

Patch-set: 7