Update patch set 14
Patch Set 14:

> Patch Set 13:
> 
> My concern is that "Auto-raise" option in Deskbar seems to imply that hovering over Deskbar will trigger it, while in fact it only does it at the screen's edge and that's why I created #15678 because it wasn't obvious to me. 

No it doesn't. It's no obvious what auto-raise does but I think that it now works as expected. We could add a tool tip that appears over the check mark "Automatically raise Deskbar when mouse is on the screen edge" or similar. 

> I still think that hovering over leaf area should trigger it.

Nope because it raises on the screen edge including the edge that the leaf is on.

Patch-set: 14
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>