Update patch set 15
Patch Set 15: Code-Review+1

> Patch Set 15:

Auto-lowering is working again with latest Patch-15.

> I don't think auto-raise should be changed to the default but the option now works a lot better. Auto-raise is better as an option, like auto-hide, some love it and some hate it.

True, but since we're all for good defaults, it'd be the right decision to do auto-rise/lower out of the box if lovers >> haters. I believe that would be the case. Shall we try a poll at the forums after this change is merged and in the nightlies?

Patch-set: 15
Reviewer: Gerrit User 1000004 <1000004@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+1