Update patch set 1
Patch Set 1:

> Yes, the userguide needs to be updated at the tool online, not here. However, I don't think information on taking bug reports even belongs in an application page in the user guide.

I think it does when there may be errors that point to look at these logs.

As I didn't read this comment before, I already added the new text to the user guide page of HaikuDepot. I simplified the text a bit and adapted it to the user guide style, proof-readings welcome:

------------
Diagnostic Logs

As HaikuDepot is used, it interacts with network as well as local services. Things can go wrong, sometimes in ways that were not anticipated when the application was written. In those cases the error message may be quite unspecific. To help diagnosing the problem, HaikuDepot can provide a log of what the application does.

An error message may suggest to view the logs to get further in-depth information or you may be asked to provide the log when you file an issue at the Haiku bug tracking system.

To see the diagnostic logs, HaikuDepot needs to be started from Terminal:

HaikuDepot -v debug

The command uses the -v flag to control the verbosity of the diagnostic logs. Possible values are: off, info, debug and trace. Using off will do only a small amount of logging, while trace does it in the highest detail.

As HaikuDepot starts up and when interacting with the application, you will see text scroll by in the Terminal window; this is the diagnostic log.

Patch-set: 1