Update patch set 1
Patch Set 1:

Hi; I think this would be really useful in the user guide because people keep getting confused about it and I can't see anywhere more appropriate to put it.

-------

I will update the PR to remove the change to the user-guide HTML and just leave the changes to the alert-messages' text.  Can I ask where one would ordinarily edit the content of the user-guide?

-------

The only correction I can see.

> To help diagnosing the problem

To help diagnosing a given problem

-------

Thanks.

Patch-set: 1