BPopUpMenu: Call MenusBeginning() and MenusEnded()
... hook methods when opening and closing a BPopUpMenu asynchronously.
Check to see if we are attached to a BMenuBar first, If we are, (e.g. in a
BMenuField) BMenuBar will call MenusBeginning() and MenuEnded() for us.

Create private static class method _HasMenuBar() for this purpose. Add a
minor optimization - if topmost menu anscestor is ourself don't even bother
trying to perform a dynamic_cast to see if we're a BMenuBar.

Update documentation for BWindow::MenusBeginning() and BWindow::MenusEnded().
Remove speculation on message constant names and mention of private constants.

Change-Id: I0161272f968db419e4b2887ddebbef90810183c4