BPopUpMenu: Call MenusBeginning() and MenusEnded() hook methods
... when opening and closing a BPopUpMenu that isn't attached to a BMenuBar
asynchronously. Check to see if we are attached to a BMenuBar first because
if we are attached to a BMenuBar (as is the case for example in a BMenuField)
the BMenuBar will call MenusBeginning() and MenuEnding() for us. Create
private static class method _HasMenuBar() for this purpose.

Update documentation for BWindow::MenusBeginning() and BWindow::MenusEnded().