Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I applied the patch but it doesn't make a difference for that Deskbar problem.

You have to rebuild your libbe.so and put it in your ~/config/non-packaged/lib/ directory for the patch to work. All this patch does is to make it so that the MenusBeginning and MenusEnded hook methods get called, it's up to the application to actually do something with them. In this case, with the patched Deskbar it should allow any replicant that pop a menu (most of them) to work like other menus in Deskbar with auto-raise.

Patch-set: 1
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>