Update patch set 1
Patch Set 1:

The PowerPC archtecture can be built both for openfirmware (Apple machines) or u-boot (modern Amiga-inspired devices, basically). The MMU code is at least partially different (in one case we use the OpenFirmware to get the initial memory map, in the other, we have to get it from the FDT instead).

So I don't think it should be moved.

The situation may be different on ARM, where you may need to use FDT in both u-boot and EFI platforms? Or does EFI provide services so that we don't need to handle the FDT directly ourselves for this?

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>