Update patch set 4
Patch Set 4:

Hi,
Thanks for working on the menus, it's been on the TODO list for way too long time.

Maybe I'm overcautious but I'm worried that this could cause regressions. Should we try to release beta2 and merge after that? Or will it be safe nough to get in beta2?

Patch-set: 4
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>