Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> What is purpose of this change? Does it match BeOS behavior?

The purpose of this change is to get replicant icons in Deskbar that pop a menu to call the window menu hook methods. As far as matching BeOS goes it is entirely possible that this bug is also present on BeOS but nobody noticed. Here's why, 99% of the time when your app pops a menu there is a BMenuBar somewhere in the process which calls the hook methods for you. Deskbar replicants are an exception to this rule, but Deskbar didn't care about these hook methods until I reimplemented auto-raise.

Patch-set: 3
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>