Update patch set 1
Patch Set 1:

I'm looking at the existing BMenu code and there is no direct way to track the parent window.
 
In many places, it will check if the Supermenu() is a BMenuBar, and if that's the case, it will send messages to it. But this does not work for context menus.

Your current approach isn't great because it requires support from the main application thread, and because it broadcasts the message to all windows, a lot of which didn't ask for anything.

A maybe less intrusive way would be to use StartWatching()/SendNotices() as a way to communicate between the window and the menu. This can be used only to send a B_OBSERVER_NOTICE_CHANGE message, but I don't see a reason for needing a real "modifiers changed" message here. The window could watch its own menus only, and get notified when the currently running menu needs an update.

But going even further, I wonder why you'd want to use the parent window at all. The menu can process messages itself, so why not handle things there directly? BMenu can be subclassed and you can override its MessageReceived method. I think you could change the menu items from there?

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>