Update patch set 17
Patch Set 17:

Problems drawing the leaf in the wrong position at different font sizes and the scroll arrows not attaching and detaching themselves properly in all cases in horizontal mode have been fixed in patch set 17. The hvif team icon has also been reworked to make it look less "fuzzy". I resized the icon in Icon-O-Matic to take up the whole box and regenerated the hvif. It's still not as sharp as the bitmap version so I kept the bitmap still if the team icon is not scaled based based on a bigger or smaller font size.

Patch-set: 17
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>