Update patch set 1
Patch Set 1:

None of the interface kit APIs are safe to call from outside the window thread without locking. It's up to the caller to manage that. In some cases we even have a debugger("looper must be locked") call.

The idea is that you would normally use these functions in one of 3 cases:
- From the window constructor (safe because the thread is not running yet, and the constructor makes sure the looper is locked already)
- From inside the thread (MessageReceived, AttachedToWindow, etc) (safe because you know the looper locked itelf before entering these functions)
- From outside the thread (all other cases) (unsafe, so you must remember to lock/unlock in that case)

The reason for keeping the lock/unlock outside the methods is that it makes the first two cases faster (we know no locking is needed there). Only the last case (which is not the recommended thing to do) needs the lock/unlock.

Is there a reason for BMenuItem to be different than other classes here, and do the locking by itself?

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>