Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> This code is wrong. SetShortcut can be called when fWindow == NULL (menu is not displayed). In this case fShortcutChar and fModifiers will be not updated.

The existing code checks that fWindow is not NULL, I only added locking. If fWindow is NULL setting a shortcut is not expected to work. fWindow is not NULL when the window is hidden. My patch may be no good but not for the reasons you listed.

> Your code is probably crash because you don't properly handle case where menu is not attached to window.

It's possible that there is a bug in my code but this is not the reason for the crash. The crash occurs when you SetShortcut many times in a row, if the problem was as you say then it should crash the first or maybe second time and be easy to spot.

Patch-set: 1
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>