Update patch set 4
Patch Set 4:

It seems like a massive improvement to the code to me. I could work on this new code, where as the old code was almost impossible to understand.

I will try to apply this patch myself and give it a try.

Thanks a lot for this work X512!

Patch-set: 4
CC: Gerrit User 1000109 <1000109@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>