Update patch set 2
Patch Set 2:

Good except use M instead of W, why?

em, em space, em quad
A common unit of measurement in typography. Em is traditionally defined as the width of the uppercase M in the current face and point size. It is more properly defined as simply the current point size. For example, in 12-point type, em is a distance of 12 points.

https://www.adobe.com/products/type/adobe-type-references-tips/glossary.html

Patch-set: 2