Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> My test program: https://github.com/X547/HaikuUtils/tree/master/MenuDynItems.
> Open "0" menu and press shift.

I downloaded your test program and I have to say that it does a better job than I thought was possible, I created a slightly updated version that sets some shortcut keys and adds some more shiftable menus (you may like it):
https://github.com/jscipione/HaikuUtils/commit/82cb2c1c1484244ab1041b71ca7632b4322bd643

Calling SetEventMask(B_KEYBOARD_EVENTS) on the menu does in-fact allow me to receive B_MODIFIERS_CHANGED messages in my menu without further work. In patch set 3 I have returned BMenuWindow::DispatchMessage() back to default and used the EventMask instead.

However your approach of removing and re-adding the BMenuItem to update the menu has a couple of drawbacks that I fixed in BMenuWindow and therefore have kept in patch set 3. The first drawback is that if there is a submenu open when you add/remove an item in the parent menu it will resize the menu but the submenu will not be informed of the change and will be in the wrong position. I have fixed this by listening for LayoutChanged() events and calling BMenuFrame::MoveSubmenusOver() recursively. The second drawback to your approach is that if you have a submenu open and the superitem label changes it will close the submenu. This is an inherit issue caused by removing and re-adding the superitem. With my approach you can change the superitem label and modifiers without submenus being closed.

Note that in order for Patch set 3 to function correctly when context clicking on a file in Tracker https://review.haiku-os.org/c/haiku/+/2289 must first be applied so that MenusBeginning and MenusEnded hook methods are called.

Patch-set: 3
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>