BMenuWindow: Use B_VERTICAL group layout
Resize the menu window to menu width on LayoutChanged(). When a menu window
is resized, move all submenus over recursively to make room for the larger
or smaller menu.

Add static member method MoveSubmenusOver() to BMenu class to do this work.
The method is static because it is called recursively.

Put menu frame and scroller arrows in the layout.

Update copyright year to 2020.

Change-Id: I1741e9f5c46e8c4541d734306f322f9fe82eb94e