BMenuWindow: Set to use B_VERTICAL group layout.
Resize the window in LayoutChanged(). When a menu window is resized,
move submenus over recursively to make room for the larger or smaler
menu.

Put scroller arrows in the layout and fix the drawing issues. When you
scroll up the bottom arrow goes away and I don't know why, it's a feature!

Change-Id: I1741e9f5c46e8c4541d734306f322f9fe82eb94e