Update patch set 2
Patch Set 2:

> Patch Set 1: Code-Review-1
> 
> Resizing window and layouting is implemented in Menu.cpp, any layouting changes should be done there. MenuWindow is just container.
The layout for window is done in MenuWindow, not BMenu. Only BMenuWindow and BMenuFrame are layed out in this file, no layout is done on BMenu.

BMenuFrame::MoveSubmenusOver() could go in BMenu as it is generally useful for laying out a bunch of BMenus. However, it is here for now because that's where it is being used. It could be a static class method on BMenu because it is called recursively that might make sense.

> I am not sure that sub menus should be moved. It is confusing and sub menus can go away of screen.

Of course they should, your code is less than ideal and you're making excuses.

Patch-set: 2
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>